refactor(program-doc): switch callers to unified DOCProgram
