+++ src/features/program-doc/__tests__/DOCProgram.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
-import DOCProgramContainer from "../DOCProgramContainer";
+import DOCProgram from "../DOCProgram";
 import { projectDatas } from "@shared/lib/data";
 
-describe("DOCProgramContainer (characterization)", () => {
+describe("DOCProgram (characterization)", () => {
     // 풍부한 데이터를 가진 프로젝트 (이미지/결과/url 있음)
     const richProject = projectDatas.find(
         (p: any) => p.projectImages?.length > 0 && p.projectReulst?.length > 0
@@ -19,7 +19,7 @@ describe("DOCProgramContainer (characterization)", () => {
 
     it("프로젝트명을 렌더한다", () => {
         render(
-            <DOCProgramContainer type="DOC" name={richProject.projectName} />
+            <DOCProgram type="DOC" name={richProject.projectName} />
         );
         expect(
             screen.getAllByText(richProject.projectName).length
@@ -28,14 +28,14 @@ describe("DOCProgramContainer (characterization)", () => {
 
     it("프로젝트 설명을 렌더한다", () => {
         render(
-            <DOCProgramContainer type="DOC" name={richProject.projectName} />
+            <DOCProgram type="DOC" name={richProject.projectName} />
         );
         expect(screen.getByText(richProject.projectDesc)).toBeInTheDocument();
     });
 
     it("프로젝트 성과 항목들을 렌더한다", () => {
         render(
-            <DOCProgramContainer type="DOC" name={richProject.projectName} />
+            <DOCProgram type="DOC" name={richProject.projectName} />
         );
         richProject.projectReulst.forEach((r: any) => {
             // 제목은 "1. 제목" 형태로 렌더되므로 부분 매칭 (여러 노드에 매칭될 수 있음)
@@ -59,7 +59,7 @@ describe("DOCProgramContainer (characterization)", () => {
 
     it("스택 아이템 이름들을 렌더한다", () => {
         render(
-            <DOCProgramContainer type="DOC" name={richProject.projectName} />
+            <DOCProgram type="DOC" name={richProject.projectName} />
         );
         richProject.stack.forEach((s: any) => {
             expect(screen.getByText(s.name)).toBeInTheDocument();
@@ -69,7 +69,7 @@ describe("DOCProgramContainer (characterization)", () => {
     it("프로젝트 이미지가 없을 때 안내 문구를 표시한다", () => {
         if (!noImageProject) return;
         render(
-            <DOCProgramContainer
+            <DOCProgram
                 type="DOC"
                 name={noImageProject.projectName}
             />
@@ -82,7 +82,7 @@ describe("DOCProgramContainer (characterization)", () => {
     it("url이 비어있을 때 '공개된 URL 없음'을 표시한다", () => {
         if (!noUrlProject) return;
         render(
-            <DOCProgramContainer type="DOC" name={noUrlProject.projectName} />
+            <DOCProgram type="DOC" name={noUrlProject.projectName} />
         );
         expect(screen.getByText("공개된 URL 없음")).toBeInTheDocument();
     });
