+++ src/features/program-doc/index.ts
@@ -1,2 +1,1 @@
-export { default as DOCProgramContainer } from "./DOCProgramContainer";
-export { default as DOCProgramComponent } from "./DOCProgramComponent";
+export { default as DOCProgram } from "./DOCProgram";
